mais pourtant elle tourne

+++ S5engine/src/core/properties/dummycontrolable.cpp
@@ -18,23 +18,23 @@ void DummyControlable::update(double elapsed_time)
 	if(manager.getButtonStatus(m_control_forward) == InputManager::Held)
 	{
 		Node* parentNode = this->node();
-		Vector3d move = parentNode->getRotation() * Vector3d(0.0, 0.0, -elapsed_time);
+		Vector3f move = parentNode->getRotation() * Vector3f(0.0, 0.0, -elapsed_time);
 		parentNode->move(move);
 	}
 	if(manager.getButtonStatus(m_control_backward) == InputManager::Held)
 	{
 		Node* parentNode = this->node();
-		Vector3d move = parentNode->getRotation() * Vector3d(0.0, 0.0, elapsed_time);
+		Vector3f move = parentNode->getRotation() * Vector3f(0.0, 0.0, elapsed_time);
 		parentNode->move(move);
 	}
 	if(manager.getButtonStatus(m_control_left) == InputManager::Held)
 	{
 		Node* parentNode = this->node();
-		parentNode->rotate(Vector3d(0.0, 1.0, 0.0), 130 * elapsed_time);
+		parentNode->rotate(Vector3f(0.0, 1.0, 0.0), 130 * elapsed_time);
 	}
 	if(manager.getButtonStatus(m_control_right) == InputManager::Held)
 	{
 		Node* parentNode = this->node();
-		parentNode->rotate(Vector3d(0.0, 1.0, 0.0), - 130 * elapsed_time);
+		parentNode->rotate(Vector3f(0.0, 1.0, 0.0), - 130 * elapsed_time);
 	}
 }
